+++ tests/src/test/spark331/scala/org/apache/spark/sql/rapids/RapidsTestsBaseTrait.scala
@@ -29,31 +29,33 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.sql
 
-import org.apache.spark.sql.types._
+/*** spark-rapids-shim-json-lines
+{"spark": "331"}
+spark-rapids-shim-json-lines ***/
+package org.apache.spark.sql.rapids
 
-object RapidsTestConstants {
+import org.apache.spark.sql.rapids.utils.BackendTestSettings
 
-  val RAPIDS_TEST: String = "Rapids - "
+trait RapidsTestsBaseTrait {
 
-  val IGNORE_ALL: String = "IGNORE_ALL"
+  protected val rootPath: String = getClass.getResource("/").getPath
+  protected val basePath: String = rootPath + "unit-tests-working-home"
 
-  val SUPPORTED_DATA_TYPE = TypeCollection(
-    BooleanType,
-    ByteType,
-    ShortType,
-    IntegerType,
-    LongType,
-    FloatType,
-    DoubleType,
-    DecimalType,
-    StringType,
-    BinaryType,
-    DateType,
-    TimestampType,
-    ArrayType,
-    StructType,
-    MapType
-  )
+  protected val warehouse: String = basePath + "/spark-warehouse"
+  protected val metaStorePathAbsolute: String = basePath + "/meta"
+
+  // The blacklist is taken in highest priority. Tests on the
+  // list will never be run with no regard to backend test settings.
+  def testNameBlackList: Seq[String] = Seq()
+
+  def shouldRun(testName: String): Boolean = {
+    if (testNameBlackList.exists(_.equalsIgnoreCase(RapidsTestConstants.IGNORE_ALL))) {
+      return false
+    }
+    if (testNameBlackList.contains(testName)) {
+      return false
+    }
+    BackendTestSettings.shouldRun(getClass.getCanonicalName, testName)
+  }
 }
