+++ tests/src/test/spark331/scala/org/apache/spark/sql/rapids/utils/BackendTestSettings.scala
@@ -29,6 +29,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+/*** spark-rapids-shim-json-lines
+{"spark": "331"}
+spark-rapids-shim-json-lines ***/
 package org.apache.spark.sql.rapids.utils
 
 import java.util
